Create staff form - sepearate form
